ALTUSDT_FUTURES/requirements.txt
# ALT/USDT Futures Trading Bot Requirements

# Core dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Logging and display
rich>=13.0.0

# Database
# sqlite3 is included in Python standard library

# Optional for enhanced features
asyncio-throttle>=1.0.0
aiohttp>=3.8.0

# Development tools (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0